requirements.txt
SpeechRecognition 
pyttsx3 
wikipedia 
requests 
Pillow